Updating styling for Landing page
