+++ src/components/VisTwoExplanation.jsx
@@ -11,7 +11,7 @@ export default class Vis3Explanation extends Component {
   return (
     <Container fluid>
       <center>
-        <Header as= 'hVis1' className='analysis-title'><u>VISUALIZATION 2</u></Header>
+        <Header as= 'hVis1' className='analysis-title'>VISUALIZATION 2</Header>
           <p className='paragraph'>
           <br/>
           Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis id tellus eu risus pretium porta. Sed venenatis turpis a arcu porttitor, sed varius elit semper. Fusce quis facilisis augue, vel finibus nisl. Phasellus suscipit semper sapien a aliquet. Fusce sed ultricies diam. Pellentesque faucibus fringilla massa vitae consequat. Aliquam porta ultrices risus, at mattis purus. Donec id orci placerat, ornare magna eu, malesuada justo. Suspendisse potenti. Donec faucibus volutpat purus, non vulputate lorem finibus sit amet.
