+++ src/pages/Landing.jsx
@@ -3,12 +3,12 @@ import { browserHistory } from 'react-router'
 import { Container, Header, Dropdown, Divider } from 'semantic-ui-react'
 
 import StateMap from './StateMap'
-import Vis1Explanation from './Vis1Explanation'
-import Vis2Explanation from './Vis2Explanation'
-import Vis3Explanation from './Vis3Explanation'
-import Visualization1 from './Visualization1'
-import Visualization2 from './Visualization2'
-import Visualization3 from './Visualization3'
+import VisOneExplanation from '../components/VisOneExplanation'
+import VisTwoExplanation from '../components/VisTwoExplanation'
+import VisThreeExplanation from '../components/VisThreeExplanation'
+import VisualizationOne from '../components/VisualizationOne'
+import VisualizationTwo from '../components/VisualizationTwo'
+import VisualizationThree from '../components/VisualizationThree'
 
 class Landing extends Component {
   constructor (props) {
@@ -59,17 +59,18 @@ class Landing extends Component {
         </Container>
 
         <Container className='analysis-section' fluid>
-          <Vis1Explanation />
-          <Visualization1 />
+          <VisOneExplanation />
+          <VisualizationOne width={1000} height={400} />
         </Container>
 
         <Container className='analysis-section' fluid>
-          <Visualization2 width={700} height={400} />
+          <VisTwoExplanation />
+          <VisualizationTwo width={1000} height={400} />
         </Container>
 
         <Container className='analysis-section' fluid>
-          <Vis3Explanation />
-          <Visualization3 />
+          <VisThreeExplanation />
+          <VisualizationThree width={1000} height={400} />
         </Container>
 
       </Container>
