+++ src/styles/index.sass
@@ -106,23 +106,30 @@ svg
     margin-top: 50px
 
 .analysis-section
-    //height: 700px
     display: flex !important
     flex-direction: column
     justify-content: center
     align-items: center
-    border-top: 10px
-    border-top-style: ridge
-    border-color: #8f052c
     padding: 20px
 
     .analysis-title
-    padding: 50px
-    color: #000
-    text-align: center
-    justify-content: center
-    align-items: center
-    font-size: 24px
+      font-weight: normal !important
+      font-family: 'Avenir Next' !important
+      font-size: 28px !important
+      color: #3F3F3F !important
+      text-transform: capitalize !important
+      margin: 0 !important
+      margin-top: 40px !important
+
+    .paragraph
+      font-weight: normal !important
+      font-family: 'Avenir Next' !important
+      text-transform: capitalize !important
+      color: #707070 !important
+      font-size: 14px !important
+      margin: 0 !important
+      margin-bottom: 20px !important
+      max-width: 1000px
 
 .paragraph
     color: #000
